requirements.txt
Flask
flask-limiter
textstat
python-dotenv
flask-cors
requests
gunicorn
google-genai
